[Fix] FilteringNode bug sampling rate
When the signal was resampled, the original sampling rate was still being used
Ref #19

+++ backend/app/main.py
@@ -116,7 +116,7 @@ async def options_filtering():
 async def filtering(
     signal: str = Form(...,
                        description="JSON-encoded list of `[timestamp, value]` pairs."),
-    sampling_rate: int = Form(...,
+    sampling_rate: float = Form(...,
                               description="Sampling rate of the input signal in Hz."),
     filter_config: str = Form(
         ..., description="JSON-encoded dict including `method`, `lowcut`, `highcut`, `order`, or `python`."),
